Fix Gradio type coercion bug in load_random_example
Fixed bug where clicking "Predict Popularity" after loading a random
example would fail with "'<' not supported between instances of 'str'
and 'int'" error. The issue was that load_random_example() was returning
raw pandas/numpy types instead of proper Python float types, causing
Gradio Radio components to handle values inconsistently.

Changes:
- Convert all numeric values to float() in load_random_example()
- Ensures consistent type handling across Gradio components
- Prevents string/int comparison errors in predict_popularity()

Also updated PowerPoint presentation with tuned model metrics:
- R² = 0.4772 (was 0.8447/0.39)
- RMSE = 16.06 (was 5.74)
- MAE = 11.95 (was 4.54)
- Estimators = 450 (was 200)

Created update_presentation_metrics.py script for automated slide updates.

+++ app_gradio.py
@@ -580,12 +580,14 @@ def load_random_example():
     """Load a random track example from the dataset"""
     random_track = df.sample(n=1).iloc[0]
 
+    # Ensure all numeric values are proper Python types (not numpy types)
+    # to avoid type coercion issues with Gradio components
     return (
         float(random_track['danceability']),
         float(random_track['energy']),
-        int(random_track['key']),
+        float(random_track['key']),  # Changed to float for Slider compatibility
         float(random_track['loudness']),
-        int(random_track['mode']),
+        float(random_track['mode']),  # Changed to float - Gradio Radio may expect this
         float(random_track['speechiness']),
         float(random_track['acousticness']),
         float(random_track['instrumentalness']),
@@ -593,7 +595,7 @@ def load_random_example():
         float(random_track['valence']),
         float(random_track['tempo']),
         float(random_track['duration_ms']),
-        int(random_track['time_signature']),
+        float(random_track['time_signature']),  # Changed to float - Gradio Radio may expect this
         str(random_track['track_genre'])
     )
